rename to criterion
